Adding DocBlock to the module * Add doc to all the API functions * compile typescript
