+++ .env
@@ -1,5 +1,5 @@
-SECRET = 
-PUBLISH = 
+SECRET = rRQ26GcsZzoEhbrP2HZvLYDbn9C9et
+PUBLISH = HGvTMLDssJghr9tlN9gr4DVYt0qyBy
 Thawani_NODE_API_DEV_URI = https://uatcheckout.thawani.om
 Thawani_NODE_API_DEV_PROD = https://checkout.thawani.om
 ENV = dev
